Move `WindowDisplaysPreview` to separate file
